+++ src/main/java/kitchenpos/products/dao/ProductDao.java
@@ -1,6 +1,6 @@
 package kitchenpos.products.dao;
 
-import kitchenpos.products.model.Product;
+import kitchenpos.products.tobe.domain.product.domain.Product;
 
 import java.util.List;
 import java.util.Optional;
