+++ src/main/java/kitchenpos/products/tobe/domain/product/domain/Product.java
@@ -1,0 +1,61 @@
+package kitchenpos.products.tobe.domain.product.domain;
+
+import java.math.BigDecimal;
+import java.util.Objects;
+
+public class Product {
+    private Long id;
+    private String name;
+    private BigDecimal price;
+
+    public Product() {}
+
+    public Product(Long id, String name, BigDecimal price) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(final Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(final String name) {
+        this.name = name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public void setPrice(final BigDecimal price) {
+        this.price = price;
+    }
+
+    public ProductDomain toDomain() {
+        return new ProductDomain(this);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Product product = (Product) o;
+        return Objects.equals(id, product.id) &&
+                Objects.equals(name, product.name) &&
+                Objects.equals(price.setScale(0), product.price);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, name, price);
+    }
+}
